Set up key metrics etc
